refactor(metadata): move API clients to metadata feature
- Move CrossRefAPI, OpenAlexAPI, SemanticScholarAPI from src/apis/ to src/features/metadata/apis/
- Update imports in moved files to use relative path for BaseMetadataAPI
- Update imports in MetadataFetcher.ts and index.ts to use new paths
- Update metadata apis index to export all four APIs
- Update test imports to use new module path
