+++ src/__tests__/unit/OpenAlexAPI.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
-import { OpenAlexAPI } from '../../apis/OpenAlexAPI';
+import { OpenAlexAPI } from '@/features/metadata/apis';
 import type { SearchQuery, OpenAlexWork } from '@/shared/core/types';
 
 // Mock Zotero HTTP
