+++ src/apis/index.ts
@@ -1,10 +1,5 @@
-export { CrossRefAPI } from './CrossRefAPI';
-export { OpenAlexAPI } from './OpenAlexAPI';
-export { SemanticScholarAPI } from './SemanticScholarAPI';
 export { LibGenAPI } from './LibGenAPI';
 export { ArxivAPI } from './ArxivAPI';
 export { PubMedCentralAPI } from './PubMedCentralAPI';
 
-// Note: Other APIs to be implemented in future phases:
-// export { LibGenAPI } from './LibGenAPI';
-// export { ArxivAPI } from './ArxivAPI'; 
+// Note: CrossRefAPI, OpenAlexAPI, SemanticScholarAPI have been moved to @/features/metadata/apis
