+++ src/features/metadata/apis/index.ts
@@ -1,1 +1,4 @@
 export { BaseMetadataAPI } from './BaseMetadataAPI';
+export { CrossRefAPI } from './CrossRefAPI';
+export { OpenAlexAPI } from './OpenAlexAPI';
+export { SemanticScholarAPI } from './SemanticScholarAPI';
